+++ lib/fun_with_flags/store/persistent/redis.ex
@@ -14,9 +14,16 @@ defmodule FunWithFlags.Store.Persistent.Redis do
 
   def worker_spec do
     import Supervisor.Spec, only: [worker: 3]
-    worker(Redix, [Config.redis_config, @conn_options], [restart: :permanent])
-  end
 
+    conf = case Config.redis_config do
+      uri when is_binary(uri) ->
+        [uri, @conn_options]
+      opts when is_list(opts) ->
+        [Keyword.merge(opts, @conn_options)]
+    end
+
+    worker(Redix, conf, [restart: :permanent])
+  end
 
   def get(flag_name) do
     case Redix.command(@conn, ["HGETALL", format(flag_name)]) do